pyproject.toml
[project]
name = "rag-notebook-studio"
version = "0.1.0"
description = "Notebook-first RAG system using LangChain, FAISS, HuggingFace embeddings, and Groq LLMs."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "flask>=3.1.3",
    "ipykernel>=7.2.0",
    "jq>=1.11.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "unstructured[docx,xlsx]>=0.21.2",
]
